LensKit version improvements
- expose version info in core from LenskitInfo
- expose head revision
- print head revision in LensKit CLI
