+++ lenskit-cli/src/main/java/org/grouplens/lenskit/cli/Main.java
@@ -25,6 +25,7 @@ import net.sourceforge.argparse4j.inf.*;
 import org.apache.commons.lang3.SystemUtils;
 import org.apache.commons.lang3.reflect.ConstructorUtils;
 import org.apache.commons.lang3.reflect.MethodUtils;
+import org.grouplens.lenskit.core.LenskitInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -61,8 +62,9 @@ public class Main {
             Logging.configureLogging(options);
             Runtime rt = Runtime.getRuntime();
             logger.info("Starting LensKit {} on Java {} from {}",
-                        Version.lenskitVersion(),
+                        LenskitInfo.lenskitVersion(),
                         SystemUtils.JAVA_VERSION, SystemUtils.JAVA_VENDOR);
+            logger.debug("Built from Git revision {}", LenskitInfo.getHeadRevision());
             logger.debug("Using VM '{}' version {} from {}",
                          SystemUtils.JAVA_VM_NAME,
                          SystemUtils.JAVA_VM_VERSION,
