+++ lenskit-core/src/main/meta/version.properties
